+++ plugin/src/main/java/io/github/eugene239/gradle/plugin/dependency/internal/LibKey.kt
@@ -1,5 +1,8 @@
 package io.github.eugene239.gradle.plugin.dependency.internal
 
+import org.gradle.api.artifacts.result.ResolvedComponentResult
+
+
 internal data class LibIdentifier(
     val group: String,
     val module: String
@@ -12,12 +15,22 @@ internal data class LibIdentifier(
 internal data class LibKey(
     val group: String,
     val module: String,
-    val version: String
+    val version: String,
 ) {
     override fun toString(): String {
         return "$group:$module:$version"
     }
 }
+
+internal data class LibDetails(
+    val key: LibKey,
+    val isStrict: Boolean,
+    val isSubmodule: Boolean,
+    val repositoryId: String?,
+    @Transient
+    // not null only if submodule
+    val result: ResolvedComponentResult? = null
+)
 
 internal fun LibKey.toIdentifier(): LibIdentifier {
     return LibIdentifier(