+++ plugin/src/main/java/io/github/eugene239/gradle/plugin/dependency/internal/StartupFlags.kt
@@ -4,5 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 internal data class StartupFlags(
-    val fetchVersions: Boolean
+    val fetchVersions: Boolean = false,
+    val fetchLibSize: Boolean = false
 )