+++ plugin/src/main/java/io/github/eugene239/gradle/plugin/dependency/internal/exception/RepositoryException.kt
@@ -5,6 +5,7 @@ import io.github.eugene239.gradle.plugin.dependency.internal.service.MavenMetada
 
 internal sealed class RepositoryException(override val message: String) : Exception(message) {
     internal data class RepositoryNotFoundException(private val lib: LibKey) : RepositoryException("Repository not found for lib: $lib")
+    internal data class MissingRepositoryName(private val lib: LibKey) : RepositoryException("Missing repository name lib: $lib")
     internal data class RepositoryWithVersionInMetadataNotFound(private val lib: LibKey, private val metadataSet: Set<MavenMetadata>) :
         RepositoryException("Repository found metadata for :$lib, but exact version not found: ${metadataSet.mapNotNull { it.url }}")
 }