+++ plugin/src/main/java/io/github/eugene239/gradle/plugin/dependency/internal/output/graph/DefaultGraphOutput.kt
@@ -1,8 +1,10 @@
 package io.github.eugene239.gradle.plugin.dependency.internal.output.graph
 
+import io.github.eugene239.gradle.plugin.dependency.internal.LibIdentifier
 import io.github.eugene239.gradle.plugin.dependency.internal.output.graph.model.FlatDependencies
 import io.github.eugene239.gradle.plugin.dependency.internal.output.graph.model.PluginConfiguration
 import io.github.eugene239.gradle.plugin.dependency.internal.output.graph.model.TopDependencies
+import io.github.eugene239.gradle.plugin.dependency.internal.Versions
 import io.github.eugene239.gradle.plugin.dependency.internal.ui.UiSaver
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.encodeToJsonElement
@@ -18,10 +20,16 @@ internal class DefaultGraphOutput(
         explicitNulls = false
     }
 
-    override fun save(pluginConfiguration: PluginConfiguration, flatDependencies: FlatDependencies, topDependencies: TopDependencies): File {
+    override fun save(
+        pluginConfiguration: PluginConfiguration,
+        flatDependencies: FlatDependencies,
+        topDependencies: TopDependencies,
+        libVersions: Map<String, Map<LibIdentifier, Versions>>
+    ): File {
         savePluginConfiguration(pluginConfiguration)
         saveFlatDependencies(flatDependencies)
         saveTopDependencies(topDependencies)
+        saveLibVersions(libVersions)
         return uiSaver.save(rootDir)
     }
 
@@ -48,6 +56,19 @@ internal class DefaultGraphOutput(
             val file = File(dir, "top-dependencies.json")
             file.createNewFile()
             val json = prettyEncoder.encodeToString(dependencies.sorted())
+            file.writeText(json)
+        }
+    }
+
+    private fun saveLibVersions(confToLibVersions: Map<String, Map<LibIdentifier, Versions>>) {
+        confToLibVersions.forEach { (configuration, conflicts) ->
+            val dir = File(rootDir, configuration)
+            dir.mkdirs()
+
+            val data = conflicts.mapKeys { entry -> entry.key.toString() }
+            val json = prettyEncoder.encodeToString(data)
+            val file = File(dir, "conflicts.json")
+            file.createNewFile()
             file.writeText(json)
         }
     }