+++ plugin/src/main/java/io/github/eugene239/gradle/plugin/dependency/internal/service/DefaultMavenService.kt
@@ -5,6 +5,7 @@ import io.github.eugene239.gradle.plugin.dependency.internal.LibKey
 import io.github.eugene239.gradle.plugin.dependency.internal.exception.PomException
 import io.github.eugene239.gradle.plugin.dependency.internal.exception.WIPException
 import io.github.eugene239.gradle.plugin.dependency.internal.provider.RepositoryProvider
+import io.github.eugene239.gradle.plugin.dependency.internal.service.simplexml.SimpleXmlConverter
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.engine.cio.CIO
@@ -15,21 +16,21 @@ import io.ktor.client.request.head
 import io.ktor.http.ContentType
 import io.ktor.http.HttpStatusCode
 import io.ktor.http.isSuccess
-import io.ktor.serialization.kotlinx.xml.xml
 import kotlinx.coroutines.sync.withPermit
 import org.gradle.api.logging.Logger
 import java.net.URL
 
 internal class DefaultMavenService(
     private val logger: Logger,
-    private val repositoryProvider: RepositoryProvider
+    private val repositoryProvider: RepositoryProvider,
+    private val timeoutMillis: Long,
 ) : MavenService {
 
     private val client: HttpClient = HttpClient(CIO) {
         install(HttpTimeout) {
-            requestTimeoutMillis = 10_000
-            connectTimeoutMillis = 10_000
-            socketTimeoutMillis = 10_000
+            requestTimeoutMillis = timeoutMillis
+            connectTimeoutMillis = timeoutMillis
+            socketTimeoutMillis = timeoutMillis
         }
 //        install(HttpRequestRetry) {
 //            retryOnExceptionIf(
@@ -44,10 +45,7 @@ internal class DefaultMavenService(
 //            delayMillis { retry: Int -> retry * 1000L }
 //        }
         install(ContentNegotiation) {
-            xml(
-                contentType = ContentType.Any,
-                format = XmlFormat.format
-            )
+            register(ContentType.Any, SimpleXmlConverter())
         }
     }