+++ plugin/src/main/java/io/github/eugene239/gradle/plugin/dependency/internal/service/XmlFormat.kt
@@ -1,18 +1,16 @@
 package io.github.eugene239.gradle.plugin.dependency.internal.service
 
-import nl.adaptivity.xmlutil.ExperimentalXmlUtilApi
-import nl.adaptivity.xmlutil.serialization.DefaultXmlSerializationPolicy
-import nl.adaptivity.xmlutil.serialization.XML
-import nl.adaptivity.xmlutil.serialization.XmlConfig
-import nl.adaptivity.xmlutil.serialization.defaultSharedFormatCache
+import io.github.eugene239.gradle.plugin.dependency.internal.service.simplexml.PropertiesConverter
+import org.simpleframework.xml.convert.Registry
+import org.simpleframework.xml.convert.RegistryStrategy
+import org.simpleframework.xml.core.Persister
 
 internal object XmlFormat {
 
-    @OptIn(ExperimentalXmlUtilApi::class)
-    val format = XML {
-        policy = DefaultXmlSerializationPolicy(formatCache = defaultSharedFormatCache()) {
-            unknownChildHandler = XmlConfig.IGNORING_UNKNOWN_CHILD_HANDLER
-        }
-        isCachingEnabled = false
-    }
+    val serializer = Persister(
+        RegistryStrategy(
+            Registry()
+                .bind(Properties::class.java, PropertiesConverter::class.java)
+        )
+    )
 }