+++ plugin/src/main/java/io/github/eugene239/gradle/plugin/dependency/task/SingleDependencyTask.kt
@@ -2,6 +2,7 @@ package io.github.eugene239.gradle.plugin.dependency.task
 
 import io.github.eugene239.gradle.plugin.dependency.internal.LibKey
 import io.github.eugene239.gradle.plugin.dependency.internal.provider.DefaultRepositoryProvider
+import io.github.eugene239.gradle.plugin.dependency.internal.service.DefaultMavenService
 import io.github.eugene239.gradle.plugin.dependency.internal.usecase.SingleDependencyUseCase
 import io.github.eugene239.gradle.plugin.dependency.internal.usecase.SingleDependencyUseCaseParams
 import org.gradle.api.tasks.Input
@@ -15,13 +16,23 @@ abstract class SingleDependencyTask : BaseTask() {
     @Option(option = "name", description = "Name of the dependency with version")
     var dependency: String = ""
 
+    @Input
+    @Option(option = "connection-timeout", description = "Ktor client connection timeout")
+    var connectionTimeout: String = "$DEFAULT_CONNECTION_TIMEOUT"
+
+    private val repositoryProvider = DefaultRepositoryProvider(
+        project = project,
+        logger = logger,
+        limit = 1
+    )
 
     private val useCase = SingleDependencyUseCase(
         logger = logger,
-        repositoryProvider = DefaultRepositoryProvider(
-            project = project,
+        repositoryProvider = repositoryProvider,
+        mavenService = DefaultMavenService(
             logger = logger,
-            limit = 1
+            repositoryProvider = repositoryProvider,
+            timeoutMillis = connectionTimeout.toLongOrNull() ?: DEFAULT_CONNECTION_TIMEOUT
         )
     )