+++ ui/src/screens/NotFoundScreen.vue
@@ -1,0 +1,20 @@
+<script>
+export default {
+  name: "NotFoundScreen"
+}
+</script>
+
+<template>
+  <section class="hero is-fullheight-with-navbar">
+    <div class="hero-body is-flex is-justify-content-center is-align-items-center">
+      <div class="is-primary">
+        <p class="title">Not Found</p>
+        <router-link class="subtitle tag is-danger is-light mt-2" to="/">Return to home</router-link>
+      </div>
+    </div>
+  </section>
+</template>
+
+<style scoped>
+
+</style>