+++ CHANGELOG.md
@@ -5,16 +5,38 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
-### Added
+#### Added
+
+#### Removed
+
+#### Fixed
+
+
+## [Released]
+
+### [0.3.0] - 2025-03-24
+#### Added
+- simplexml
+- conflicts.json
+- lib details to track strict version and submodule libraries
+
+#### Removed
+- kotlinx-serialization-xml
+
+#### Fixed
+- reduced execution time
+- logic for multi module project
+- resolved to latest
+
+
+### [0.2.0] - 2025-03-09
+#### Added
 - Template to format markdown
 - Local maven repo for `demolib`
 
-### Removed
+#### Removed
 - Unused and deprecated classes
 - Koin
 
-### Fixed
-
-
-## [0.1.0] - 2025-02-06
+### [0.1.0] - 2025-02-06
 - Initial release